Lift meta-to-config side effect from Gallery to App
The effect that pushes `meta.cdn` / `defaultTheme` / etc. into the
SPA's `config` singleton lived inside the Gallery component. When
an admin landed directly on /m/... (or refreshed the page there),
Gallery never mounted, the effect never ran, and config.PHOTO_ROOT_URL
stayed at its default `/` instead of the operator's configured CDN
URL — so thumbnail / display image requests went to the wrong host
and failed.

Move the same effect into App.tsx, which mounts on every route.
react-query's cache key (["meta"]) is shared with Gallery's own
metaQuery, so it's still one network call regardless of which
route booted first.

Tidy up the now-unused beta-store imports in Gallery/index.tsx.
