+++ react-app/src/App.tsx
@@ -7,6 +7,7 @@ import {
 } from "react-router-dom";
 import styled from "@emotion/styled";
 import { useTranslation } from "react-i18next";
+import { useQuery } from "@tanstack/react-query";
 
 import "./App.css";
 
@@ -22,7 +23,21 @@ import Notifications from "./components/Notifications";
 import LoginModal from "./components/LoginModal";
 import ChangePasswordModal from "./components/ChangePasswordModal";
 
-import { useLangStore } from "./stores";
+import config from "./lib/config";
+import metaService from "./services/meta";
+import { useBetaStore, useLangStore } from "./stores";
+import { BETA_FEATURES, type BetaFeature, type BetaMode } from "./stores/beta";
+
+interface Meta {
+  cdn?: string;
+  name?: string;
+  description?: string;
+  defaultGallery?: string;
+  defaultTheme?: string;
+  initialGalleryView?: string;
+  firstWeekday?: string | number;
+  betaFeatures?: Record<string, string>;
+}
 
 const Footer = styled.div`
   width: 100%;
@@ -41,6 +56,37 @@ const App = (): React.ReactElement => {
   // subviews and rendering with `undefined` would cause runtime errors. The
   // lang store loads the active language's dictionary at module load.
   const countryDataReady = useLangStore((s) => s.countryData !== undefined);
+
+  // Boot-side: fetch /meta and push the runtime-tunable bits into the
+  // SPA's config singleton. Has to live here (not in Gallery) so the
+  // Manage routes get correct PHOTO_ROOT_URL etc. when an admin lands
+  // directly on /m/... — Gallery doesn't mount on those routes. The
+  // react-query cache key is shared with Gallery's own ["meta"]
+  // query, so it's still one network call.
+  const metaQuery = useQuery({
+    queryKey: ["meta"],
+    queryFn: () => metaService.getAll(),
+  });
+  const meta = metaQuery.data as Meta | undefined;
+  const setBetaModes = useBetaStore((s) => s.setModes);
+  React.useEffect(() => {
+    if (!meta) return;
+    config.PHOTO_ROOT_URL = meta.cdn || config.PHOTO_ROOT_URL;
+    if (meta.defaultGallery) config.DEFAULT_GALLERY = meta.defaultGallery;
+    if (meta.defaultTheme) config.DEFAULT_THEME = meta.defaultTheme;
+    if (meta.initialGalleryView)
+      config.INITIAL_GALLERY_VIEW = meta.initialGalleryView;
+    if (meta.firstWeekday !== undefined)
+      config.FIRST_WEEKDAY = Number(meta.firstWeekday);
+    if (meta.betaFeatures) {
+      const next: Partial<Record<BetaFeature, BetaMode>> = {};
+      for (const f of BETA_FEATURES) {
+        const v = meta.betaFeatures[f];
+        if (v === "on" || v === "off" || v === "user") next[f] = v;
+      }
+      if (Object.keys(next).length > 0) setBetaModes(next);
+    }
+  }, [meta, setBetaModes]);
 
   return (
     <>
